+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log (@egomobile/http-server)
 
+## 0.2.0
+
+- add `params()` middleware
+- `next()` function of a middleware can receive an error object now
+- make sure, that all handlers and middlewares are async functions now
+- update type descriptions for handlers and middlewars
+- (bug-)fixes
+
 ## 0.1.2
 
 - initial release